module-2-docker/exercise-2-mc_potts_file_io/requirements.txt
matplotlib==3.5.3
numpy==1.23.2
PyYAML==6.0
